Crate flite

FLITE / FRAME-LITE

Polkadot-SDK’s FRAME, but low on the dose of complexity, therefore lite.

This is a an experimental library meant only for demonstration purposes, and is not meant for production use. Yet, it might influence the decision making to those who want to build reusable FRAME-based libraries.

Put differently, this crate demonstrates a number of techniques that can be used to build more opinionated, yet simpler variants of FRAME.

Techniques

The main techniques used in this library are as follows:

NOT Relying on frame_system (and cumulus::parachain_system).

FRAME comes with an underrated macro: #[pallet::disable_frame_system_supertrait_check]. This macro enables a pallet to rely on a pallet other than frame_system. This opens the door for a team to write an “opinionated flavor” of frame_system for their own use.

In this crate, we introduce flite_system, an opinionated alternative to [polkadot_sdk::frame_system].

Constrained Associated Types

Constrained associated types are a powerful and an underrated feature of Rust, and they are crucially important for FRAME.

FRAME expresses the majority of its flexibility through associated types. Constrained associated types are then therefore a way for one to express limitations on those associated types.

The downside of this would be that the users would have less flexibility. Yet, the benefit would be that the syntax would be less generic, and more concrete.

We argue that for a starter FRAME template, and for beginner users it is much easier to work with concrete types for common concepts like AccountId and Balance, rather than generic variants.

trait Currency {
    type Balance;

    fn get() -> Balance {};
}

/// A pallet that is not constraining `Currency::Balance`, and must always refer to it as
trait Config {
    type Currency: Currency

    fn usage() {
        // This pallet doesn't really now what `<Self::Currency as Currency>::Balance` is.
        let _: <Self::Currency as Currency>::Balance = T::Currency::get();
    }
}

/// A pallet that is constraining `Currency::Balance`, to JUST `u128`.
trait Config {
    type Currency: Currency<Balance = u128>;

    fn usage() {
        // This pallet knows that `Self::Currency::Balance` is `u128`.
        let _: u128 = T::Currency::get();
    }
}

flite_system constraints a number of important associated types, to be seen in the example further below.

Tight Coupling To Our Benefit

Generally, tight coupling is considered a bad thing. Yet, in the context of a library like flite, we can use it to our advantage. If we tightly couple flite_system with a number of pallets, we achieve:

We can express that a runtime can only work if flite_system and the tightly coupled pallets are present in it.

We can safely provide a set of default for all the configurations of the tightly coupled pallets.

We can also express constrains on the configurations as constraints on the associated types, as explained above.

Example: flite_system

With the above context in mind, let’s look and analyze flite_system::Config, which puts all of the above practices in one place.

#[pallet::config]
pub trait Config:
	polkadot_sdk::frame_system::Config<AccountId = types::AccountId>
	+ psdk::pallet_timestamp::Config
	+ psdk::pallet_aura::Config
	+ psdk::pallet_balances::Config<
		Balance = types::Balance,
		RuntimeHoldReason = u8,
		RuntimeFreezeReason = u8,
	>
{
}

Notice how flite is constraining the AccountId to be types::AccountId, the Balance to be types::Balance, and even more advance things like HoldReason.

This means flite_system can only be used in a runtime with such constants. More nicely, a pallet using flite_system can already be written with these concrete types in mind.

Simplifying a New Pallet

Let’s see how now we can simplify the process of writing a new pallet using that relies on flite_system.

First, this pallet is indeed declared as such:

#[pallet::config]
#[pallet::disable_frame_system_supertrait_check]
pub trait Config: flite::flite_system::Config {}

This then allows the pallet to be written with concrete types in mind:

Currency Simplification

We also present two ideas on how a simpler currency system can be built on top of the above abstractions. These are discussed in native_currency.

First, a pallet can import and use FRAME’s currency abstractions in the old school way. That is, pull the currency abstractions via trait Config:

#[pallet::config]
#[pallet::disable_frame_system_supertrait_check]
pub trait Config: flite::flite_system::Config {
	type AdvanceCurrency: flite::native_currency::Advance;
}

And then use them as such:

pub fn advance_currency_usage(
	origin: OriginFor<T>,
	to: flite::types::AccountId,
	amount: flite::types::Balance,
) -> DispatchResult {
	use frame::traits::fungible::*;
	let who: flite::types::AccountId = ensure_signed(origin)?;

	// read
	let _: flite::types::Balance = T::AdvanceCurrency::balance(&who);
	// transfer
	T::AdvanceCurrency::transfer(
		&who,
		&to,
		amount,
		frame::traits::tokens::Preservation::Expendable,
	)?;
	// 'hold'.
	T::AdvanceCurrency::hold(&0u8, &who, amount)?;

	Ok(())
}

Note that this already comes with some improvements. This code so far looks just like the normal FRAME code, except there is no trace of the notorious BalanceOf type. Moreover, there is no special consideration needed for RuntimeHoldReason conversion. RuntimeHoldReason is simply a u8. All of this is because flite_system has already constrained these types.

flite can also provide even simpler, even more opinionated currency abstractions, such as native_currency::Simple. The usage of such a simple abstraction can look like:

pub fn simple_currency_usage(
	origin: OriginFor<T>,
	to: flite::types::AccountId,
	amount: u128, // aka. flite::types::Balance
) -> DispatchResult {
	let who: flite::types::AccountId = ensure_signed(origin)?;

	// read. Yes, the balance types is just `u128`
	let _: u128 = flite::native_currency::Simple::<T>::balance_of(&who);
	// transfer
	flite::native_currency::Simple::<T>::transfer(&who, &to, amount)?;
	// 'lock', which is an abstraction over holding with more opinions
	flite::native_currency::Simple::<T>::lock(&who, amount)?;

	Ok(())
}

A Step Further: Defaults For Tightly Coupled Pallets

As the final step, flite could also simplify the process of runtime configuration, if it could provide defaults for all the pallets that it is tightly coupling itself with.

All of these defaults can be audited and guaranteed to work nicely together, as long as none of them is tweaked by the end user. If a mild degree of flexibility is needed, we instead provide a separate, higher level trait that can be used to parameterize all other pallets with.

In default_configs, we provide one such example. The end-user entry-point is merely default_configs::FliteConfigurations. This is one trait that we expect the end user to parameterize for us. Instead of a lot of the technical nitty-gritty details that [polkadot_sdk::frame_system] exposes, this trait is meant to be very easy to comprehend to a novice developer.

Then, based on FliteConfigurations, we provide defaults for each pallet’s Config traits:

default_configs::FliteFrameSystem

default_configs::FliteTimestamp

default_configs::FliteAura

default_configs::FliteBalances

These defaults are provided in a format that should be compatible with FRAME’s native derive_impl (or the underlying macro_magic crate), yet this part is not yet working in this demo and therefore the runtime file contains a lot of boilerplate.

Notice how all of these types are generic over F, which is a FliteConfigurations.

This technique allows us to retain some degree fo flexibility, and abstract over complicated relationships. For example, (almost) all runtimes have [polkadot_sdk::pallet_timestamp::Config::MinimumPeriod] set to half of [polkadot_sdk::pallet_aura::Config::SlotDuration]. We abstract over this needless complication by providing one simple configuration to the user FliteConfigurations::BlockTime (which is what we mean by SlotDuration). Under the hood, we take care of providing the correct defaults to each of the aforementioned pallets.

Dependencies

This work is dependent on the following PR(s)

https://github.com/paritytech/polkadot-sdk/pull/5246

Further Ideas

Create abstractions over [polkadot_sdk::pallet_assets], and the [polkadot_sdk::polkadot_sdk_frame::traits::fungibles] as well, for example hardcoding the AssetId to be u32.

More pallets that are reasonable always needed: authorship, session, etc.

flite, and flite-parachain. If one has to be chosen, we care more about parachain.

flite can provide a single set of weights that are safe to use in weights.

flite_system::Pallet can provide further standalone helper functions, such as flite_system::Pallet::block_author.

integrity_test in flite_system is super important, as it is the one place where we can check the configuration of all pallets. For example, the relationship between BlockTime, SlotDuration and MinimumPeriod can be expressed here. The ability to check these within a pallet is yet another benefit of tight coupling.

Reflection

Everything here is backwards compatible, and a developer who is advance can always fallback into the ADVANCE-FRAME-land.

Re-exports

pub use default_configs::FliteConfigurations;

Modules

default_configs

Default configurations for the pallets that flite_system::Config is tightly coupled with.

flite_system

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

native_currency

types

The opinionated types that we dictate.

weights

Flite provides a set of weights for its users. Technically, this should be something like a well-benchmarked, audited set of weights that are safe to use.