Trait flite::default_configs::FliteAuraDefaultConfig

pub trait FliteAuraDefaultConfig {
    type AuthorityId;
    type MaxAuthorities;
    type DisabledValidators;
    type AllowMultipleBlocksPerSlot;
    type SlotDuration;
}

A flite representative of [polkadot_sdk::pallet_aura::Config].

This merely exists because pallet_aura itself doesn’t have #[pallet::config(with_default)]. Although, as we established for FliteTimestamp, even if it would, it would have not been useful for this case.

We drop all most trait bounds, as we test it and we are professional.

Required Associated Types

type AuthorityId

See [polkadot_sdk::pallet_aura::Config::AuthorityId].

type MaxAuthorities

See [polkadot_sdk::pallet_aura::Config::MaxAuthorities].

type DisabledValidators

See [polkadot_sdk::pallet_aura::Config::DisabledValidators].

type AllowMultipleBlocksPerSlot

See [polkadot_sdk::pallet_aura::Config::AllowMultipleBlocksPerSlot].

type SlotDuration

See [polkadot_sdk::pallet_aura::Config::SlotDuration].