Trait flite::flite_system::Config

pub trait Config:
    Config<AccountId = AccountId>
    + Config
    + Config
    + Config<Balance = Balance, RuntimeHoldReason = u8, RuntimeFreezeReason = u8> { }

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait. Flite puts a lot of restrictions on what types are fed to account Id. All of this is fine as long as the users use the provided [config_preludes::FliteFrameSystem].

Object Safety

This trait is not object safe.