flite::native_currency

Trait Advance

pub trait Advance:
    Inspect<AccountId, Balance = Balance>
    + Mutate<AccountId>
    + InspectHold<AccountId, Reason = u8>
    + MutateHold<AccountId> { }

Trait to give you access to the advance underlying FRAME currency primitives. You can also import all of these manually into your trait, this is merely a syntactic sugar.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.